pyproject.toml
[project]
name = "prabhanshu-website"
version = "0.1.0"
description = "Personal website built with FastHTML"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml>=0.7.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.hatch.build.targets.wheel]
packages = ["website"]
